kayttaja-luokan lisäys

+++ Budjetointisovellus/src/main/java/budjetointisovellus/domain/BudgetService.java
@@ -1,23 +1,28 @@
 package budjetointisovellus.domain;
 
 import budjetointisovellus.dao.BudgetDao;
+import budjetointisovellus.dao.UserDao;
 import budjetointisovellus.domain.Budget;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 /**
- * Sovelluslogiikasta vastaava luokka, joka välittää pyynnöt tietokantaa 
+ * Sovelluslogiikasta vastaava luokka, joka välittää pyynnöt tietokantaa
  * käsittelevälle rajapinnalle
  */
 public class BudgetService {
 
     private BudgetDao budgetDao;
+    private UserDao userDao;
+    private User loggedIn;
 
-    public BudgetService(BudgetDao budgetDao) {
+    public BudgetService(BudgetDao budgetDao, UserDao userDao) {
         this.budgetDao = budgetDao;
+        this.userDao = userDao;
     }
 
     /**
@@ -25,8 +30,8 @@ public class BudgetService {
      *
      * @name budjettia kuvaava nimi käyttäjän syöttämässä muodossa.
      */
-    public boolean create(String name) {
-        Budget budget = new Budget(name);
+    public boolean createBudget(String name) {
+        Budget budget = new Budget(name, loggedIn);
 
         try {
             budgetDao.create(budget);
@@ -36,15 +41,45 @@ public class BudgetService {
         return true;
     }
 
+    public boolean createUser(String name) {
+        if (userDao.findByUsername(name) != null) {
+            return false;
+        }
+
+        User user = new User(name);
+
+        try {
+            userDao.create(user);
+        } catch (Exception ex) {
+            return false;
+        }
+        return true;
+    }
+
+    public boolean logIn(String username) {
+        User user = userDao.findByUsername(username);
+        if (user == null) {
+            return false;
+        }
+
+        this.loggedIn = user;
+
+        return true;
+    }
+
+    public void logOut() {
+        this.loggedIn = null;
+    }
+
     /**
      * Pyytää rajapintaa poistamaan budjetin.
-     * 
+     *
      * @budgetName käyttäjän syöttämä budjetin nimi.
      */
     public boolean removeBudget(String budgetName) {
         try {
             for (Budget one : budgetDao.getBudgets()) {
-                if (one.getName().equals(budgetName)) {
+                if (one.getName().equals(budgetName) && one.getUser().equals(loggedIn)) {
                     budgetDao.remove(one);
                     return true;
                 }
@@ -57,25 +92,28 @@ public class BudgetService {
 
     /**
      * Pyytää rajapinnalta budjetit.
-     * 
+     *
      * @return palauttaa haetut budjetit.
      */
     public List<Budget> getBudgets() {
-        return budgetDao.getBudgets();
+        return budgetDao.getBudgets()
+                .stream()
+                .filter(t -> t.getUser().equals(loggedIn))
+                .collect(Collectors.toList());
     }
 
     /**
      * Pyytää rajapintaa lisäämään erän tiettyyn budjettiin.
-     * 
+     *
      * @budget käyttäjän syöttämä budjetin nimi.
      * @name käyttäjän syöttämä erän nimi.
      * @amount käyttäjän syöttämä lukumäärä erälle.
      */
-    public boolean addEventToBudget(String budget, String actionName, int amount) {
+    public boolean addTransactionToBudget(String budget, String actionName, int amount) {
         try {
             for (Budget b : budgetDao.getBudgets()) {
-                if (b.getName().equals(budget)) {
-                    budgetDao.create(b, actionName, amount);
+                if (b.getName().equals(budget) && b.getUser().equals(loggedIn)) {
+                    budgetDao.createTransaction(b, actionName, amount);
                     return true;
                 }
             }
@@ -87,7 +125,7 @@ public class BudgetService {
 
     /**
      * Pyytää rajapinnalta tiettyyn budjettiin liittyvät erät.
-     * 
+     *
      * @name käyttäjän syöttämä budjetin nimi.
      * @return saadut erät listana.
      */
@@ -95,24 +133,24 @@ public class BudgetService {
         ArrayList<Transaction> list = new ArrayList<>();
 
         for (Budget budget : budgetDao.getBudgets()) {
-            if (budget.getName().equals(name)) {
-                list = budget.getEvents();
+            if (budget.getName().equals(name) && budget.getUser().equals(loggedIn)) {
+                list = budget.getTransactions();
             }
         }
         return list;
     }
 
-     /**
+    /**
      * Pyytää rajapintaa poistamaan budjetin tietyn erän.
-     * 
+     *
      * @budgetName käyttäjän syöttämä budjetin nimi.
      * @name käyttäjän syöttämä erän nimi.
      */
     public boolean removeBudgetEvent(String budgetName, String name) {
         try {
             for (Budget one : budgetDao.getBudgets()) {
-                if (one.getName().equals(budgetName)) {
-                    budgetDao.remove(one, name);
+                if (one.getName().equals(budgetName) && one.getUser().equals(loggedIn)) {
+                    budgetDao.removeTransaction(one, name);
                 }
             }
         } catch (Exception ex) {
